Initial refactoring stage.py to support multiple training algorithms
